agents/trend-scout/requirements.txt
aiohttp==3.9.1
feedparser==6.0.11
pytrends==4.9.2
redis==5.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
pydantic==2.5.3
python-dateutil==2.8.2
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0